@@ -1,5 +1,8 @@
 /*
- * (c) Copyright 2016 Palantir Technologies Inc. All rights reserved.
+ * (c) Copyright 2016-2017 Palantir Technologies Inc. All rights reserved.
+ *
+ * (c) Copyright 2023 brightSPARK Labs (from commit `c2774cac049bb0007d14790527ea2499670fef83` onwards).
+ * All rights reserved.
  */
 
 package com.palantir.websecurity.examples;
@@ -10,28 +13,27 @@ import com.palantir.websecurity.CorsConfiguration;
 import com.palantir.websecurity.WebSecurityBundle;
 import com.palantir.websecurity.WebSecurityConfigurable;
 import com.palantir.websecurity.WebSecurityConfiguration;
-import io.dropwizard.Application;
-import io.dropwizard.Configuration;
 import io.dropwizard.assets.AssetsBundle;
-import io.dropwizard.setup.Bootstrap;
-import io.dropwizard.setup.Environment;
-import javax.validation.Valid;
-import javax.validation.constraints.NotNull;
-import javax.ws.rs.GET;
-import javax.ws.rs.Path;
-import javax.ws.rs.Produces;
-import javax.ws.rs.core.MediaType;
+import io.dropwizard.core.Application;
+import io.dropwizard.core.Configuration;
+import io.dropwizard.core.setup.Bootstrap;
+import io.dropwizard.core.setup.Environment;
+import jakarta.validation.Valid;
+import jakarta.validation.constraints.NotNull;
+import jakarta.ws.rs.GET;
+import jakarta.ws.rs.Path;
+import jakarta.ws.rs.Produces;
+import jakarta.ws.rs.core.MediaType;
 
-/**
- * Example applications for ETE testing. Contains both Web + REST application examples.
- */
+/** Example applications for ETE testing. Contains both Web + REST application examples. */
 @SuppressWarnings("checkstyle:designforextension")
 public final class Example {
 
     public static final String EXAMPLES_RESOURCE_RESPONSE = "hello world";
     public static final String CSP_FROM_APP = "csp from app default";
     public static final String CTO_FROM_APP = "cto should be overridden";
-    public static final String CTO_FROM_YML = "cto is overridden yay!"; // appears in the example YAML file
+    public static final String CTO_FROM_YML =
+            "cto is overridden yay!"; // appears in the example YAML file
 
     private Example() {
         // utility class for testing
@@ -41,9 +43,7 @@ public final class Example {
         new ExampleWebApplication().run(args);
     }
 
-    /**
-     * Jersey resource used in both the Web + REST example applications.
-     */
+    /** Jersey resource used in both the Web + REST example applications. */
     @Path("hello")
     public static final class ExampleResource {
 
@@ -54,10 +54,9 @@ public final class Example {
         }
     }
 
-    /**
-     * Configuration class used in both the Web + REST example applications.
-     */
-    public static final class ExampleConfiguration extends Configuration implements WebSecurityConfigurable {
+    /** Configuration class used in both the Web + REST example applications. */
+    public static final class ExampleConfiguration extends Configuration
+            implements WebSecurityConfigurable {
 
         @JsonProperty("webSecurity")
         @NotNull
@@ -71,26 +70,26 @@ public final class Example {
     }
 
     /**
-     * Example Web Application. Sets application defaults using {@link #webSecurityDefaults}. The application is
-     * configured to serve assets from `/*` and Jersey API requests from `/api/*`.
+     * Example Web Application. Sets application defaults using {@link #webSecurityDefaults}. The
+     * application is configured to serve assets from `/*` and Jersey API requests from `/api/*`.
      */
     public static class ExampleWebApplication extends Application<ExampleConfiguration> {
 
-        private final WebSecurityConfiguration webSecurityDefaults = WebSecurityConfiguration.builder()
+        private final WebSecurityConfiguration webSecurityDefaults =
+                WebSecurityConfiguration.builder()
 
-                // set app defaults for different header values
-                .contentSecurityPolicy(CSP_FROM_APP)
-                .contentTypeOptions(CTO_FROM_APP)
+                        // set app defaults for different header values
+                        .contentSecurityPolicy(CSP_FROM_APP)
+                        .contentTypeOptions(CTO_FROM_APP)
 
-                // CORS is still DISABLED, since the allowedOrigins is not set, but the default value will be
-                // respected if it's ever turned on
-                .cors(CorsConfiguration.builder()
-                        .preflightMaxAge(60 * 10)
-                        .build())
+                        // CORS is still DISABLED, since the allowedOrigins is not set, but the
+                        // default value will be
+                        // respected if it's ever turned on
+                        .cors(CorsConfiguration.builder().preflightMaxAge(60 * 10).build())
+                        .build();
 
-                .build();
-
-        private final WebSecurityBundle webSecurityBundle = new WebSecurityBundle(this.webSecurityDefaults);
+        private final WebSecurityBundle webSecurityBundle =
+                new WebSecurityBundle(this.webSecurityDefaults);
 
         @Override
         public void initialize(Bootstrap<ExampleConfiguration> bootstrap) {
@@ -99,7 +98,8 @@ public final class Example {
         }
 
         @Override
-        public void run(ExampleConfiguration configuration, Environment environment) throws Exception {
+        public void run(ExampleConfiguration configuration, Environment environment)
+                throws Exception {
             environment.jersey().register(new ExampleResource());
         }
 
@@ -110,8 +110,9 @@ public final class Example {
     }
 
     /**
-     * Example REST Application. Uses the application defaults using {@link ExampleWebApplication#webSecurityDefaults}.
-     * The application is configured to serve Jersey API requests from `/`.
+     * Example REST Application. Uses the application defaults using {@link
+     * ExampleWebApplication#webSecurityDefaults}. The application is configured to serve Jersey API
+     * requests from `/`.
      */
     public static final class ExampleRestApplication extends ExampleWebApplication {
 
