@@ -1,64 +1,80 @@
 /*
- * (c) Copyright 2016 Palantir Technologies Inc. All rights reserved.
+ * (c) Copyright 2016-2017 Palantir Technologies Inc. All rights reserved.
+ *
+ * (c) Copyright 2023 brightSPARK Labs (from commit `c2774cac049bb0007d14790527ea2499670fef83` onwards).
+ * All rights reserved.
  */
 
 package com.palantir.websecurity.examples;
 
-import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertNull;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertNull;
 
 import com.google.common.net.HttpHeaders;
 import io.dropwizard.client.JerseyClientBuilder;
-import io.dropwizard.testing.junit.DropwizardAppRule;
-import javax.ws.rs.client.Client;
-import javax.ws.rs.core.Response;
+import io.dropwizard.testing.junit5.DropwizardAppExtension;
+import io.dropwizard.testing.junit5.DropwizardExtensionsSupport;
+import jakarta.ws.rs.client.Client;
+import jakarta.ws.rs.core.Response;
 import org.glassfish.jersey.client.ClientProperties;
-import org.junit.BeforeClass;
-import org.junit.ClassRule;
-import org.junit.Test;
+import org.junit.jupiter.api.BeforeAll;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
 
-/**
- * Tests for {@link Example.ExampleRestApplication}.
- */
+/** Tests for {@link Example.ExampleRestApplication}. */
 public final class ExampleRestTests {
 
     public static final String ORIGIN_VALUE = "http://origin.com";
 
-    @ClassRule
-    public static final DropwizardAppRule<Example.ExampleConfiguration> RULE = new DropwizardAppRule<>(
-            Example.ExampleRestApplication.class,
-            Example.ExampleRestApplication.class.getClassLoader().getResource("example-rest.yml").getPath());
+    @ExtendWith(DropwizardExtensionsSupport.class)
+    public static final DropwizardAppExtension<Example.ExampleConfiguration> RULE =
+            new DropwizardAppExtension<>(
+                    Example.ExampleRestApplication.class,
+                    Example.ExampleRestApplication.class
+                            .getClassLoader()
+                            .getResource("example-rest.yml")
+                            .getPath());
 
     private static Client client;
 
-    @BeforeClass
-    public static void beforeClass() {
-        client = new JerseyClientBuilder(RULE.getEnvironment())
-                .withProperty(ClientProperties.CONNECT_TIMEOUT, 10000)
-                .withProperty(ClientProperties.READ_TIMEOUT, 10000)
-                .build("tests");
+    @BeforeAll
+    public static void BeforeAll() {
+        client =
+                new JerseyClientBuilder(RULE.getEnvironment())
+                        .withProperty(ClientProperties.CONNECT_TIMEOUT, 10000)
+                        .withProperty(ClientProperties.READ_TIMEOUT, 10000)
+                        .build("tests");
     }
 
     @Test
     public void testCorsHeadersAppliedToApi() {
-        Response response = client
-                .target(String.format("http://localhost:%d/example-context/hello", RULE.getLocalPort())).request()
-                .header(HttpHeaders.ORIGIN, ORIGIN_VALUE)
-                .get();
+        Response response =
+                client.target(
+                                String.format(
+                                        "http://localhost:%d/example-context/hello",
+                                        RULE.getLocalPort()))
+                        .request()
+                        .header(HttpHeaders.ORIGIN, ORIGIN_VALUE)
+                        .get();
 
         // check basic functionality
         assertEquals(200, response.getStatus());
         assertEquals(Example.EXAMPLES_RESOURCE_RESPONSE, response.readEntity(String.class));
 
         // check for a YAML defined CORS entry
-        assertEquals(ORIGIN_VALUE, response.getHeaderString(HttpHeaders.ACCESS_CONTROL_ALLOW_ORIGIN));
+        assertEquals(
+                ORIGIN_VALUE, response.getHeaderString(HttpHeaders.ACCESS_CONTROL_ALLOW_ORIGIN));
     }
 
     @Test
     public void testWebSecurityHeadersNotAppliedToApi() {
-        Response response = client
-                .target(String.format("http://localhost:%d/example-context/hello", RULE.getLocalPort())).request()
-                .get();
+        Response response =
+                client.target(
+                                String.format(
+                                        "http://localhost:%d/example-context/hello",
+                                        RULE.getLocalPort()))
+                        .request()
+                        .get();
 
         // check basic functionality
         assertEquals(200, response.getStatus());
