@@ -1,21 +1,22 @@
 /*
- * (c) Copyright 2016 Palantir Technologies Inc. All rights reserved.
+ * (c) Copyright 2016-2017 Palantir Technologies Inc. All rights reserved.
+ *
+ * (c) Copyright 2023 brightSPARK Labs (from commit `c2774cac049bb0007d14790527ea2499670fef83` onwards).
+ * All rights reserved.
  */
 
 package com.palantir.websecurity.filters;
 
-import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertNull;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertNull;
 
 import com.google.common.net.HttpHeaders;
 import com.palantir.websecurity.WebSecurityConfiguration;
-import org.junit.Test;
+import org.junit.jupiter.api.Test;
 import org.springframework.mock.web.MockHttpServletRequest;
 import org.springframework.mock.web.MockHttpServletResponse;
 
-/**
- * Tests for {@link WebSecurityHeaderInjector}.
- */
+/** Tests for {@link WebSecurityHeaderInjector}. */
 public final class WebSecurityHeaderInjectorTests {
 
     private static final String USER_AGENT_NOT_IE = "not-ie-10-or-11";
@@ -26,17 +27,20 @@ public final class WebSecurityHeaderInjectorTests {
 
     @Test
     public void testDisabledNoHeaders() {
-        WebSecurityHeaderInjector injector = new WebSecurityHeaderInjector(WebSecurityConfiguration.builder()
-                .contentSecurityPolicy(WebSecurityConfiguration.TURN_OFF)
-                .contentTypeOptions(WebSecurityConfiguration.TURN_OFF)
-                .frameOptions(WebSecurityConfiguration.TURN_OFF)
-                .xssProtection(WebSecurityConfiguration.TURN_OFF)
-                .build());
+        WebSecurityHeaderInjector injector =
+                new WebSecurityHeaderInjector(
+                        WebSecurityConfiguration.builder()
+                                .contentSecurityPolicy(WebSecurityConfiguration.TURN_OFF)
+                                .contentTypeOptions(WebSecurityConfiguration.TURN_OFF)
+                                .frameOptions(WebSecurityConfiguration.TURN_OFF)
+                                .xssProtection(WebSecurityConfiguration.TURN_OFF)
+                                .build());
 
         injector.injectHeaders(request, response);
 
         assertNull(response.getHeader(HttpHeaders.CONTENT_SECURITY_POLICY));
-        assertNull(response.getHeader(WebSecurityHeaderInjector.HEADER_IE_X_CONTENT_SECURITY_POLICY));
+        assertNull(
+                response.getHeader(WebSecurityHeaderInjector.HEADER_IE_X_CONTENT_SECURITY_POLICY));
         assertNull(response.getHeader(HttpHeaders.X_CONTENT_TYPE_OPTIONS));
         assertNull(response.getHeader(HttpHeaders.X_FRAME_OPTIONS));
         assertNull(response.getHeader(HttpHeaders.X_XSS_PROTECTION));
@@ -50,7 +54,8 @@ public final class WebSecurityHeaderInjectorTests {
         request.addHeader(HttpHeaders.USER_AGENT, WebSecurityHeaderInjector.USER_AGENT_IE_10);
 
         response.addHeader(HttpHeaders.CONTENT_SECURITY_POLICY, TEST_VALUE);
-        response.addHeader(WebSecurityHeaderInjector.HEADER_IE_X_CONTENT_SECURITY_POLICY, TEST_VALUE);
+        response.addHeader(
+                WebSecurityHeaderInjector.HEADER_IE_X_CONTENT_SECURITY_POLICY, TEST_VALUE);
         response.addHeader(HttpHeaders.X_CONTENT_TYPE_OPTIONS, TEST_VALUE);
         response.addHeader(HttpHeaders.X_FRAME_OPTIONS, TEST_VALUE);
         response.addHeader(HttpHeaders.X_XSS_PROTECTION, TEST_VALUE);
@@ -58,7 +63,10 @@ public final class WebSecurityHeaderInjectorTests {
         injector.injectHeaders(request, response);
 
         assertEquals(1, response.getHeaders(HttpHeaders.CONTENT_SECURITY_POLICY).size());
-        assertEquals(1, response.getHeaders(WebSecurityHeaderInjector.HEADER_IE_X_CONTENT_SECURITY_POLICY).size());
+        assertEquals(
+                1,
+                response.getHeaders(WebSecurityHeaderInjector.HEADER_IE_X_CONTENT_SECURITY_POLICY)
+                        .size());
         assertEquals(1, response.getHeaders(HttpHeaders.X_CONTENT_TYPE_OPTIONS).size());
         assertEquals(1, response.getHeaders(HttpHeaders.X_FRAME_OPTIONS).size());
         assertEquals(1, response.getHeaders(HttpHeaders.X_XSS_PROTECTION).size());
@@ -66,51 +74,62 @@ public final class WebSecurityHeaderInjectorTests {
 
     @Test
     public void testContentSecurityPolicyNonIe10or11() {
-        WebSecurityHeaderInjector injector = new WebSecurityHeaderInjector(WebSecurityConfiguration.builder()
-                .contentSecurityPolicy(TEST_VALUE)
-                .build());
+        WebSecurityHeaderInjector injector =
+                new WebSecurityHeaderInjector(
+                        WebSecurityConfiguration.builder()
+                                .contentSecurityPolicy(TEST_VALUE)
+                                .build());
 
         request.addHeader(HttpHeaders.USER_AGENT, USER_AGENT_NOT_IE);
 
         injector.injectHeaders(request, response);
 
         assertEquals(TEST_VALUE, response.getHeader(HttpHeaders.CONTENT_SECURITY_POLICY));
-        assertNull(response.getHeader(WebSecurityHeaderInjector.HEADER_IE_X_CONTENT_SECURITY_POLICY));
+        assertNull(
+                response.getHeader(WebSecurityHeaderInjector.HEADER_IE_X_CONTENT_SECURITY_POLICY));
     }
 
     @Test
     public void testContentSecurityPolicyIe10() {
-        WebSecurityHeaderInjector injector = new WebSecurityHeaderInjector(WebSecurityConfiguration.builder()
-                .contentSecurityPolicy(TEST_VALUE)
-                .build());
+        WebSecurityHeaderInjector injector =
+                new WebSecurityHeaderInjector(
+                        WebSecurityConfiguration.builder()
+                                .contentSecurityPolicy(TEST_VALUE)
+                                .build());
 
         request.addHeader(HttpHeaders.USER_AGENT, WebSecurityHeaderInjector.USER_AGENT_IE_10);
 
         injector.injectHeaders(request, response);
 
         assertEquals(TEST_VALUE, response.getHeader(HttpHeaders.CONTENT_SECURITY_POLICY));
-        assertEquals(TEST_VALUE, response.getHeader(WebSecurityHeaderInjector.HEADER_IE_X_CONTENT_SECURITY_POLICY));
+        assertEquals(
+                TEST_VALUE,
+                response.getHeader(WebSecurityHeaderInjector.HEADER_IE_X_CONTENT_SECURITY_POLICY));
     }
 
     @Test
     public void testContentSecurityPolicyIe11() {
-        WebSecurityHeaderInjector injector = new WebSecurityHeaderInjector(WebSecurityConfiguration.builder()
-                .contentSecurityPolicy(TEST_VALUE)
-                .build());
+        WebSecurityHeaderInjector injector =
+                new WebSecurityHeaderInjector(
+                        WebSecurityConfiguration.builder()
+                                .contentSecurityPolicy(TEST_VALUE)
+                                .build());
 
         request.addHeader(HttpHeaders.USER_AGENT, WebSecurityHeaderInjector.USER_AGENT_IE_11);
 
         injector.injectHeaders(request, response);
 
         assertEquals(TEST_VALUE, response.getHeader(HttpHeaders.CONTENT_SECURITY_POLICY));
-        assertEquals(TEST_VALUE, response.getHeader(WebSecurityHeaderInjector.HEADER_IE_X_CONTENT_SECURITY_POLICY));
+        assertEquals(
+                TEST_VALUE,
+                response.getHeader(WebSecurityHeaderInjector.HEADER_IE_X_CONTENT_SECURITY_POLICY));
     }
 
     @Test
     public void testContentTypeOptions() {
-        WebSecurityHeaderInjector injector = new WebSecurityHeaderInjector(WebSecurityConfiguration.builder()
-                .contentTypeOptions(TEST_VALUE)
-                .build());
+        WebSecurityHeaderInjector injector =
+                new WebSecurityHeaderInjector(
+                        WebSecurityConfiguration.builder().contentTypeOptions(TEST_VALUE).build());
 
         injector.injectHeaders(request, response);
 
@@ -119,9 +138,9 @@ public final class WebSecurityHeaderInjectorTests {
 
     @Test
     public void testFrameOptions() {
-        WebSecurityHeaderInjector injector = new WebSecurityHeaderInjector(WebSecurityConfiguration.builder()
-                .frameOptions(TEST_VALUE)
-                .build());
+        WebSecurityHeaderInjector injector =
+                new WebSecurityHeaderInjector(
+                        WebSecurityConfiguration.builder().frameOptions(TEST_VALUE).build());
 
         injector.injectHeaders(request, response);
 
@@ -130,9 +149,9 @@ public final class WebSecurityHeaderInjectorTests {
 
     @Test
     public void testXssProtection() {
-        WebSecurityHeaderInjector injector = new WebSecurityHeaderInjector(WebSecurityConfiguration.builder()
-                .xssProtection(TEST_VALUE)
-                .build());
+        WebSecurityHeaderInjector injector =
+                new WebSecurityHeaderInjector(
+                        WebSecurityConfiguration.builder().xssProtection(TEST_VALUE).build());
 
         injector.injectHeaders(request, response);
 
