@@ -1,5 +1,8 @@
 /*
- * (c) Copyright 2016 Palantir Technologies Inc. All rights reserved.
+ * (c) Copyright 2016-2017 Palantir Technologies Inc. All rights reserved.
+ *
+ * (c) Copyright 2023 brightSPARK Labs (from commit `c2774cac049bb0007d14790527ea2499670fef83` onwards).
+ * All rights reserved.
  */
 
 package com.palantir.websecurity;
@@ -8,9 +11,7 @@ import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
 import com.google.common.base.Optional;
 import org.immutables.value.Value.Immutable;
 
-/**
- * Root-level Configuration for the {@link WebSecurityBundle}.
- */
+/** Root-level Configuration for the {@link WebSecurityBundle}. */
 @Immutable
 @ImmutableStyles
 @JsonDeserialize(as = ImmutableWebSecurityConfiguration.class)
@@ -19,34 +20,35 @@ public abstract class WebSecurityConfiguration {
     public static final String TURN_OFF = "";
 
     /**
-     * Value to be returned in the response header {@link com.google.common.net.HttpHeaders#CONTENT_SECURITY_POLICY}.
+     * Value to be returned in the response header {@link
+     * com.google.common.net.HttpHeaders#CONTENT_SECURITY_POLICY}.
      */
     public abstract Optional<String> contentSecurityPolicy();
 
     /**
-     * Value to be returned in the response header {@link com.google.common.net.HttpHeaders#X_CONTENT_TYPE_OPTIONS}.
+     * Value to be returned in the response header {@link
+     * com.google.common.net.HttpHeaders#X_CONTENT_TYPE_OPTIONS}.
      */
     public abstract Optional<String> contentTypeOptions();
 
     /**
-     * Value to be returned in the response header {@link com.google.common.net.HttpHeaders#X_FRAME_OPTIONS}.
+     * Value to be returned in the response header {@link
+     * com.google.common.net.HttpHeaders#X_FRAME_OPTIONS}.
      */
     public abstract Optional<String> frameOptions();
 
     /**
-     * Value to be returned in the response header {@link com.google.common.net.HttpHeaders#X_XSS_PROTECTION}.
+     * Value to be returned in the response header {@link
+     * com.google.common.net.HttpHeaders#X_XSS_PROTECTION}.
      */
     public abstract Optional<String> xssProtection();
 
-    /**
-     * Configuration for CORS functionality.
-     */
+    /** Configuration for CORS functionality. */
     public abstract Optional<CorsConfiguration> cors();
 
-    /**
-     * Provides a configuration with default values.
-     */
-    public static final WebSecurityConfiguration DEFAULT = WebSecurityConfiguration.builder().build();
+    /** Provides a configuration with default values. */
+    public static final WebSecurityConfiguration DEFAULT =
+            WebSecurityConfiguration.builder().build();
 
     // hides implementation details
     public static Builder builder() {
