@@ -1,5 +1,8 @@
 /*
- * (c) Copyright 2016 Palantir Technologies Inc. All rights reserved.
+ * (c) Copyright 2016-2017 Palantir Technologies Inc. All rights reserved.
+ *
+ * (c) Copyright 2023 brightSPARK Labs (from commit `c2774cac049bb0007d14790527ea2499670fef83` onwards).
+ * All rights reserved.
  */
 
 package com.palantir.websecurity.filters;
@@ -7,19 +10,20 @@ package com.palantir.websecurity.filters;
 import static com.google.common.base.Preconditions.checkNotNull;
 
 import com.palantir.websecurity.WebSecurityConfiguration;
+import jakarta.servlet.Filter;
+import jakarta.servlet.FilterChain;
+import jakarta.servlet.FilterConfig;
+import jakarta.servlet.ServletException;
+import jakarta.servlet.ServletRequest;
+import jakarta.servlet.ServletResponse;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
 import java.io.IOException;
-import javax.servlet.Filter;
-import javax.servlet.FilterChain;
-import javax.servlet.FilterConfig;
-import javax.servlet.ServletException;
-import javax.servlet.ServletRequest;
-import javax.servlet.ServletResponse;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
+import java.util.Locale;
 
 /**
- * A filter that injects the App Security headers using a {@link WebSecurityHeaderInjector} to all requests except for
- * those on the {@link #jerseyRoot} path.
+ * A filter that injects the App Security headers using a {@link WebSecurityHeaderInjector} to all
+ * requests except for those on the {@link #jerseyRoot} path.
  */
 public final class JerseyAwareWebSecurityFilter implements Filter {
 
@@ -64,13 +68,12 @@ public final class JerseyAwareWebSecurityFilter implements Filter {
     }
 
     private boolean isJerseyRequest(HttpServletRequest request) {
-        String cleanedServletPath = cleanJerseyRoot(request.getServletPath().toLowerCase());
+        String cleanedServletPath =
+                cleanJerseyRoot(request.getServletPath().toLowerCase(Locale.getDefault()));
         return this.jerseyRoot.equals(cleanedServletPath);
     }
 
-    /**
-     * Cleans the Jersey root path to start with a slash and end without a star or slash.
-     */
+    /** Cleans the Jersey root path to start with a slash and end without a star or slash. */
     private static String cleanJerseyRoot(String rawJerseyRoot) {
         String cleaned = rawJerseyRoot;
 
