@@ -1,5 +1,8 @@
 /*
- * (c) Copyright 2016 Palantir Technologies Inc. All rights reserved.
+ * (c) Copyright 2016-2017 Palantir Technologies Inc. All rights reserved.
+ *
+ * (c) Copyright 2023 brightSPARK Labs (from commit `c2774cac049bb0007d14790527ea2499670fef83` onwards).
+ * All rights reserved.
  */
 
 package com.palantir.websecurity.filters;
@@ -7,18 +10,19 @@ package com.palantir.websecurity.filters;
 import static com.google.common.base.Preconditions.checkNotNull;
 
 import com.palantir.websecurity.WebSecurityConfiguration;
+import jakarta.servlet.Filter;
+import jakarta.servlet.FilterChain;
+import jakarta.servlet.FilterConfig;
+import jakarta.servlet.ServletException;
+import jakarta.servlet.ServletRequest;
+import jakarta.servlet.ServletResponse;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
 import java.io.IOException;
-import javax.servlet.Filter;
-import javax.servlet.FilterChain;
-import javax.servlet.FilterConfig;
-import javax.servlet.ServletException;
-import javax.servlet.ServletRequest;
-import javax.servlet.ServletResponse;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
 
 /**
- * A filter that injects the App Security headers using a {@link WebSecurityHeaderInjector} to every request.
+ * A filter that injects the App Security headers using a {@link WebSecurityHeaderInjector} to every
+ * request.
  */
 public final class WebSecurityFilter implements Filter {
 
@@ -49,7 +53,8 @@ public final class WebSecurityFilter implements Filter {
         checkNotNull(chain);
 
         if (request instanceof HttpServletRequest && response instanceof HttpServletResponse) {
-            this.injector.injectHeaders((HttpServletRequest) request, (HttpServletResponse) response);
+            this.injector.injectHeaders(
+                    (HttpServletRequest) request, (HttpServletResponse) response);
         }
 
         chain.doFilter(request, response);
